refactor: group initial table planning with table creation
