+++ crates/jet3/src/bootstrap_table_create.rs
@@ -439,6 +439,49 @@ impl<'a> PlannedCreate<'a> {
     }
 }
 
+/// Composes sequential table creates with their initial rows. EXP-0087 supplies
+/// later-create page roles; EXP-0065 supplies row append placement. Combining
+/// the existing per-table writers is a candidate construction.
+pub(crate) fn compose_database_with_table_rows(
+    requests: &[crate::TableRows<'_>],
+    budget: &mut ResourceBudget,
+) -> Result<WholeFileImagePlan, ComposeError> {
+    if requests.len() > MAX_OBSERVED_TABLES {
+        return Err(ComposeError::UnobservedTableCount {
+            count: requests.len(),
+            observed: MAX_OBSERVED_TABLES,
+        });
+    }
+    budget.charge_allocation(ByteCount::new(
+        (requests.len() * size_of::<PlannedCreate<'_>>()) as u64,
+    ))?;
+    let mut creates = Vec::new();
+    creates
+        .try_reserve_exact(requests.len())
+        .map_err(|_| Error::Io {
+            operation: "reserve initial table plans",
+            kind: std::io::ErrorKind::OutOfMemory,
+        })?;
+    let mut next_page = EMPTY_DATABASE_PAGE_COUNT;
+    for (position, request) in requests.iter().enumerate() {
+        budget.charge_items(1)?;
+        if let Some(first) = requests[..position]
+            .iter()
+            .position(|earlier| earlier.table.name.eq_ignore_ascii_case(request.table.name))
+        {
+            return Err(ComposeError::DuplicateTableName {
+                first,
+                second: position,
+            });
+        }
+        let planned = PlannedCreate::new(&request.table, next_page, position == 0)?
+            .with_rows(request.rows, budget)?;
+        next_page = planned.page_count();
+        creates.push(planned);
+    }
+    compose_planned_creates(&creates, budget)
+}
+
 /// Returns the ordinals of the columns that own long-value map groups.
 fn long_value_columns<'s>(spec: &'s TableSpec<'_>) -> impl Iterator<Item = u16> + 's {
     spec.columns
